Split the submission package into reviewer and submission
