+++ server/src/modules/reviewer/submission.module.ts
@@ -1,12 +1,11 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { SubmissionsResolver } from './submissions.resolver';
-import { Submission } from './domain/submission.entity';
-import { SubmissionService } from './use-cases/submission.service';
-import { GqlAuthGuard } from '../auth/graphql.guard';
+import { Submission } from '../../packages/submission/submission.entity';
+import { SubmissionService } from './submission.service';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Submission]), GqlAuthGuard],
+  imports: [TypeOrmModule.forFeature([Submission])],
   providers: [SubmissionService, SubmissionsResolver],
 })
