+++ server/src/packages/submission/index.ts
@@ -1,0 +1,4 @@
+// submission package
+// There should be
+// - [ ] Repo
+// - [x] Entity
